refactored code file structures
