--- src/components/Card.css
@@ -1,4 +1,0 @@
-.card {
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.25);
-    border-radius: 12px;
-}
