+++ src/components/Expenses/ExpenseDate.css
@@ -12,6 +12,7 @@
 }
 
 .expense-date__month {
+  font-size: 0.75rem;
   font-weight: bold;
 }
