+++ src/components/Expenses/ExpenseItem.js
@@ -1,7 +1,7 @@
 import './ExpenseItem.css';
 import ExpenseDate from "./ExpenseDate";
 
-import Card from './Card';
+import Card from '../UI/Card';
 
 const ExpenseItem = props => {
   return (
