+++ src/components/Expenses/Expenses.js
@@ -1,6 +1,6 @@
 import "./Expenses.css";
 import ExpenseItem from "./ExpenseItem";
-import Card from "./Card";
+import Card from "../UI/Card";
 
 const Expenses = props => {
   return (
