+++ src/index.css
@@ -12,4 +12,3 @@ body {
   margin: 0;
   background-color: #3f3f3f;
 }
-
